Create openVariablesView function in console-output
